+++ src/utils/index.js
@@ -1,6 +1,7 @@
 export { default as asyncDelay } from './asyncDelay';
 export { default as cloneArray } from './cloneArray';
 export { default as cloneDeep } from './cloneDeep';
+export { default as freezeDeep } from './freezeDeep';
 export { default as generateHash } from './generateHash';
 export { default as isEqual } from './isEqual';
 export { default as kebabToCamel } from './kebabToCamel';
@@ -12,6 +13,7 @@ export { default as pascalToKebab } from './pascalToKebab';
 export { default as percentOfInteger } from './percentOfInteger';
 export { default as randomArrayElement } from './randomArrayElement';
 export { default as randomInteger } from './randomInteger';
+export { default as removeDomElement } from './removeDomElement';
 export { default as scrollTo } from './scrollTo';
 export { default as shuffle } from './shuffle';
 export { default as throttle } from './throttle';