+++ src/utils/logger/Logger.js
@@ -1,154 +1,193 @@
 /* eslint-disable no-console */
-import format from 'date-fns/format';
 import {
-  isBoolean,
   isString,
+  isArray,
   isPlainObject,
   isFunction,
 } from '../inspect';
-import {
-  DEFAULT_CONFIG,
-  LEVELS,
-  METHODS_MAP,
-  STYLES,
-} from './constants';
+import { DEFAULT_CONFIG, LEVELS } from './constants';
 import LogModel from './LogModel';
 
 export default class Logger {
   /**
    * @param {Object} [config]
+   * @param {String} [config.scope] Область применения
+   * @param {Array} [config.tags] Список тегов после тега области (первая строка лога)
    * @param {Function} [config.accessHandler] Обработчик с условием доступа
-   * @param {String} [config.scope] Область видимости экземпляра
-   * @param {String} [config.prefix] Полный префикс лога (значение начального тега)
-   * @param {Boolean} [config.showTime] Флаг для отображения времени лога
    */
   constructor(config = {}) {
+    this.options = { ...DEFAULT_CONFIG };
+
+    const isValid = Logger.validateConfig(config);
+
+    if (isValid) {
+      this.options = { ...this.options, ...config };
+    }
+  }
+
+  /**
+   * @param {Object} config
+   * @returns {Boolean}
+   * @private
+   */
+  static validateConfig(config) {
     const error = (...args) => {
-      console.error('[Logger]: constructor()\n', ...args);
+      console.error('[Logger] validateConfig()\n', ...args);
     };
 
     if (!isPlainObject(config)) {
-      error(
-        '| The "config" property is invalid!\n',
-        '| Given value:', config,
-      );
-      return;
+      error('The "config" property is invalid!\n', '| Given value:', config);
+      return false;
     }
 
-    const {
-      accessHandler,
-      scope,
-      prefix,
-      showTime,
-    } = config;
+    const { scope, tags, accessHandler } = config;
 
-    if (accessHandler && !isFunction(accessHandler)) {
-      error(
-        '| The "accessHandler" option is invalid!\n',
-        '| Given value:', accessHandler,
-      );
-      return;
-    }
     if (scope && !isString(scope)) {
-      error(
-        '| The "scope" option is invalid!\n',
-        '| Given value:', scope,
-      );
-      return;
+      error('The "scope" option is invalid!\n', '| Given value:', scope);
+      return false;
     }
-    if (prefix && !isString(prefix)) {
-      error(
-        '| The "prefix" option is invalid!\n',
-        '| Given value:', prefix,
-      );
-      return;
+    if (tags && !isArray(tags)) {
+      error('The "tags" option is invalid!\n', '| Given value:', tags);
+      return false;
     }
-    if (showTime && !isBoolean(showTime)) {
-      error(
-        '| The "showTime" option is invalid!\n',
-        '| Given value:', showTime,
-      );
-      return;
+    if (accessHandler && !isFunction(accessHandler)) {
+      error('The "accessHandler" option is invalid!\n', '| Given value:', accessHandler);
+      return false;
     }
 
-    this.options = {
-      ...DEFAULT_CONFIG,
-      ...config,
-    };
+    return true;
   }
 
+  /**
+   * @param {...*} args
+   * @returns {void}
+   */
   log(...args) {
     this.output(LEVELS.LOG, ...args);
   }
 
+  /**
+   * @param {...*} args
+   * @returns {void}
+   */
   info(...args) {
     this.output(LEVELS.INFO, ...args);
   }
 
+  /**
+   * @param {...*} args
+   * @returns {void}
+   */
   warn(...args) {
     this.output(LEVELS.WARN, ...args);
   }
 
+  /**
+   * @param {...*} args
+   * @returns {void}
+   */
   error(...args) {
     this.output(LEVELS.ERROR, ...args);
   }
 
   /**
+   * @param {Array} args
+   * @returns {Boolean}
+   * @private
+   */
+  static checkModelOnlyOne(args) {
+    const models = args.filter((argument) => argument instanceof LogModel);
+
+    return models.length <= 1;
+  }
+
+  /**
+   * @param {Array} args
+   * @returns {LogModel | null}
+   * @private
+   */
+  static getModelFromArgs(args) {
+    return args.find((argument) => argument instanceof LogModel) || null;
+  }
+
+  /**
+   * @param {String} level
+   * @param {Array} args
+   * @returns {Object}
+   * @private
+   */
+  static prepareData(level, args) {
+    let templateOutput = '';
+    let substringsOutput = [];
+    let isLogModelLast = false;
+
+    /**
+     * @param {String} template
+     * @param {Array} substrings
+     * @returns {void}
+     */
+    const setOutputData = (template, substrings) => {
+      templateOutput += template;
+      substringsOutput = [...substringsOutput, ...substrings];
+    };
+
+    args.forEach((argument) => {
+      const isLogModel = argument instanceof LogModel;
+
+      if (isLogModel) {
+        const { template, substrings } = argument.getOutputData(level);
+
+        if (!template) return;
+
+        setOutputData(template, substrings);
+        isLogModelLast = isLogModel;
+      } else {
+        const newLine = isLogModelLast ? ['\n'] : [];
+        const payload = [...newLine, argument];
+        setOutputData('', payload);
+      }
+    });
+
+    return {
+      template: templateOutput,
+      substrings: substringsOutput,
+    };
+  }
+
+  /**
+   * @param {String} level
+   * @param {...*} args
+   * @returns {void}
    * @private
    */
   output(level, ...args) {
     if (!this.hasAccess()) return;
 
-    const method = METHODS_MAP[level];
-    let { prefix } = this.options;
+    const argsData = [...args];
 
-    if (!prefix) {
-      const prefixFirstLetter = level.charAt(0).toLocaleUpperCase();
-      const prefixDefault = `${prefixFirstLetter}${level.substr(1).toLocaleLowerCase()}`;
-      let { scope } = this.options;
-      scope = scope ? `:${scope}` : '';
-
-      prefix = `${prefixDefault}${scope}`;
+    if (!Logger.checkModelOnlyOne(argsData)) {
+      console.error('[Logger] output()\n', '| Only one "LogModel" can be transferred');
+      return;
     }
 
-    let template = '%c%s%c';
-    let substrings = [
-      STYLES.TAG[level],
-      prefix,
-      null,
-    ];
+    const scope = this.options.scope ? `:${this.options.scope}` : '';
+    const tagScope = `${level}${scope}`;
+    const tags = [tagScope, ...(this.options.tags || [])];
+    const logModel = Logger.getModelFromArgs(argsData);
 
-    if (this.options.showTime) {
-      const timeNow = format(new Date(), 'HH:mm:ss.sss');
-
-      template += ' %c%s%c';
-      substrings = [
-        ...substrings,
-        STYLES.TAG.TIME,
-        timeNow,
-        null,
-      ];
+    if (logModel) {
+      logModel.injectLoggerData(tags);
     }
 
-    let isLogModelLast = false;
+    const { template, substrings } = Logger.prepareData(level, argsData);
 
-    [...args].forEach((argument) => {
-      const isLogModel = argument instanceof LogModel;
-      const newLine = isLogModelLast ? ['\n'] : [];
-      const payload = (isLogModel)
-        ? argument.getOutputData(level).substrings
-        : [...newLine, argument];
-
-      if (isLogModel) {
-        template += ` ${argument.getOutputData(level).template}`;
-      }
-      substrings = [...substrings, ...payload];
-      isLogModelLast = isLogModel;
-    });
-
-    console[method](template, ...substrings);
+    console[level](template, ...substrings);
   }
 
+  /**
+   * @returns {Boolean}
+   * @private
+   */
   hasAccess() {
     return this.options?.accessHandler();
   }